CI: put all decorators in dedicated file
